+++ silo-core/common/SiloCoreContracts.sol
@@ -19,6 +19,9 @@ library SiloCoreContracts {
     string public constant SHARE_DEBT_TOKEN = "ShareDebtToken.sol";
     string public constant SILO_LENS = "SiloLens.sol";
     string public constant SILO_ROUTER = "SiloRouter.sol";
+    string public constant INCENTIVES_CONTROLLER_FACTORY = "SiloIncentivesControllerFactory.sol";
+    string public constant INCENTIVES_CONTROLLER_GAUGE_LIKE_FACTORY
+        = "SiloIncentivesControllerGaugeLikeFactory.sol";
 }
 
 /// @notice SiloCoreDeployments library