+++ silo-vaults/contracts/libraries/ConstantsLib.sol
@@ -2,8 +2,9 @@
 pragma solidity ^0.8.28;
 
 /// @title ConstantsLib
-/// @author Morpho Labs
-/// @custom:contact security@morpho.org
+/// @dev Forked with gratitude from Morpho Labs.
+/// @author Silo Labs
+/// @custom:contact security@silo.finance
 /// @notice Library exposing constants.
 library ConstantsLib {
     /// @dev The maximum delay of a timelock.
@@ -13,7 +14,7 @@ library ConstantsLib {
     uint256 internal constant MIN_TIMELOCK = 1 minutes;
 
     /// @dev The maximum number of markets in the supply/withdraw queue.
-    uint256 internal constant MAX_QUEUE_LENGTH = 30;
+    uint256 internal constant MAX_QUEUE_LENGTH = 5; // TODO revert after debuggind
 
     /// @dev The maximum fee the vault can have (50%).
     uint256 internal constant MAX_FEE = 0.5e18;