+++ silo-vaults/test/foundry/ReallocateIdleTest.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.8.28;
 
 import {SiloMathLib} from "silo-core/contracts/lib/SiloMathLib.sol";
 
-import {MarketAllocation} from "../../contracts/interfaces/IMetaMorpho.sol";
+import {MarketAllocation} from "../../contracts/interfaces/ISiloVault.sol";
 
 import {IntegrationTest} from "./helpers/IntegrationTest.sol";